Refactor dashboard management pages to use new layout and components; replaced existing code with ManageMenu, ManageOffers, and ManageUsers components for improved structure and maintainability.
