+++ src/index.css
@@ -2,3 +2,6 @@
 /* @import "tailwindcss/base";
 @import "tailwindcss/components";
 @import "tailwindcss/utilities"; */
+:root{
+    --primary-orange: #ff6900;
+}
